REFACTOR: fixed app imports
