+++ src/routes/auth.route.ts
@@ -1,9 +1,9 @@
 import { Router } from 'express'
 import loginLimiter from '../middleware/login-limiter'
 import { createSessionHandler, createUserHandler, destroySessionHandler, refreshTokenHandler } from '../controllers/auth.controller'
-import validateInput from '../middleware/validateInput'
+import validateInput from '../middleware/validate-input'
 import { createSessionSchema, createUserSchema } from '../schema/user.schema'
-import requireUser from '../middleware/requireuser'
+import requireUser from '../middleware/require-user'
 
 const router = Router()
